rename lists.ts to objectUtils.ts
