+++ src/app/features/editor/utils/objectUtils.ts
